Moved useSetLanguageFromUrl into hooks
